Original Build of the Data Creation Project (WIP)
This upload is the original build of my current project (currently considered "Data Creation", though the name is likely to change). It is of course a Work-In-Progress, as there are many issues to be sorted through with the project itself, be it the issues are about the user interface, runtime efficiency, test cases, documentation, etc,. With time, such issues should be addressed.