Data_Creation1.1.5.py added
The new version-in-progress now allows the user to select a CSV file to upload through a Tkinter GUI. The upload should ideally be through the trajectory. From there, though the interface only stops at being able to select the main three variables, it will soon allow the proper data selection. From there, it should be easy to create a way to go from one Tkinter GUI to another (from the dataset upload to the ability to select variables to compare to being able to select comparisons through the Tkinter Treeview).