Data Creation Version 1.1.9 Added
Version 1.1.9 now allows the user to select a specific CSV dataset and then to select the variables (specifically, 3 variables that are one-column or multi-column, at the moment). This version has many possibilities for growth as there are test cases to be had, as well as the potential for being able to dictate the amount of columns.
The "One-column" vs "Multi-column" option may be gotten rid of to favor "Multi-column" only, though that would require an overhaul of the original 1.0.0 version, since that relies on the options being specified such that Tkinter Treeview can be utilized in a specific manner. Of course, the data can still make such an accomplishment if its unordered nature can still be utilized.