Data Creation Version 2.0 Added
The new version is now quite close to being able to implement the data to send it from the first tkinter GUI to Tkinter Treeview. The current problem is that the data format has to be decided on: categorical variable columns and data columns can be selected from the data itself, though the data misses the ability to properly separate categorical variables and data columns (currently, each can be selected, though technically are optional, which poses an obvious logistical problem).