Data ETL Version 2.2.1 Added
The project is known as a sort of ETL (currently labeled "Data ETL"). The new version still isn't ready to further the ETL steps, though it showcases ideas that had been brainstormed. Though the documentation is not fully formed for the current version, it has included the idea of combining recursion and iteration. It is difficult to implement as of now because the levels of the statistical tests that can be performed have to be decided on (such difference has a striking resemblance to the multi-way interaction problems faced when using 3 or more variables).