Version 2.2.6 Added
The new version is now closer to bridging the gap through the second interface. One specific note: the second interface attempts the capability to display what option one is selecting (conceptually speaking) to be "fully-displayed" through Treeview. The display (as with all of the interface) ends up generally limited, though that should be improved with time. There is an error such that the display overlaps rather than simply changing from one display set to another.