Uploaded Version 2.2.11
The new version experiments with treeview insertions. The difficulty: nested for-loops alone are not enough; execution functions can suffice, though the errors are not diagnosable as of yet. Recursion using stacks is also considered; efficiency is all the while a concern.
Note: due to other projects, this new version of Point_and_Click_ETL had been placed on hold about a week.